requirements.txt
flask
flask-cors
gspread
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
python-dotenv
gunicorn
pandas
openpyxl
pytz
reportlab
flask-sqlalchemy
sqlalchemy
psycopg2-binary